refactor: move export functions to a separate class
